Various minor refactorings
