+++ cli/src/main/java/de/jplag/cli/options/LanguageConverter.java
@@ -1,4 +1,4 @@
-package de.jplag.cli;
+package de.jplag.cli.options;
 
 import de.jplag.Language;
