+++ cli/src/main/java/de/jplag/cli/picocli/HelpFactory.java
@@ -1,4 +1,4 @@
-package de.jplag.cli;
+package de.jplag.cli.picocli;
 
 import picocli.CommandLine;
