+++ cli/src/main/java/de/jplag/cli/server/Routing.java
@@ -3,7 +3,7 @@ package de.jplag.cli.server;
 import com.sun.net.httpserver.HttpExchange;
 
 /**
- * Handles the data for a url prefix.
+ * Handles the data for an url prefix.
  */
 public interface Routing {
     /**
@@ -15,7 +15,7 @@ public interface Routing {
 
     /**
      * Gets the data for the given url
-     * @param subPath The remaining suffix of the url, that is not jet interpreted
+     * @param subPath The remaining suffix of the url, that is not yet interpreted
      * @param request The original http request
      * @param viewer The current report viewer
      * @return The data to respond with
