+++ cli/src/test/java/de/jplag/cli/CustomHelpTests.java
@@ -1,5 +1,8 @@
 package de.jplag.cli;
 
+import de.jplag.cli.picocli.CustomHelp;
+import de.jplag.cli.picocli.HelpFactory;
+import de.jplag.cli.picocli.ParamLabelRenderer;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -26,7 +29,6 @@ class CustomHelpTests {
      */
     @Test
     void testReturnsCustomRenderer() {
-        Assertions.assertTrue(this.help.parameterLabelRenderer() instanceof ParamLabelRenderer,
-                "The custom help object returned the wrong ParamLabelRenderer type.");
+        Assertions.assertInstanceOf(ParamLabelRenderer.class, this.help.parameterLabelRenderer(), "The custom help object returned the wrong ParamLabelRenderer type.");
     }
 }
