+++ cli/src/test/java/de/jplag/cli/LanguageTest.java
@@ -5,6 +5,8 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import java.util.Arrays;
 import java.util.List;
 
+import de.jplag.cli.options.CliOptions;
+import de.jplag.cli.options.LanguageLoader;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
